✨ Add hero component for Rome inspiration
