+++ src/app/page.tsx
@@ -1,11 +1,12 @@
 // src/app/page.tsx
 
-import { WelcomeForm } from '@/features/home';
+import { Hero, PlanForm } from '@/features/home';
 
 export default function Home() {
   return (
     <main id="main-content" className="space-y-16">
-      <WelcomeForm />
+      <Hero />
+      <PlanForm />
     </main>
   );
 }
