+++ src/features/home/components/index.ts
@@ -1,6 +1,7 @@
 // src/features/home/components/index.ts
 
-export { default as WelcomeForm } from './WelcomeForm';
+export { default as Hero } from './Hero';
+export { default as PlanForm } from './PlanForm';
 export { default as FeaturePreview } from './FeaturePreview';
 export { default as DestinationInput } from './DestinationInput';
 export { default as InspirationLink } from './InspirationLink';
